Remove old BucketTree
